requirements.txt
tensorflow
pandas
numpy
matplotlib
io
ibm_boto3
ibm_botocore.client